+++ infrastructure/database/models/auto_delete_message.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer
+from sqlalchemy import Integer, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import BaseModel
@@ -7,6 +7,6 @@ from .base import BaseModel
 class AutoDeleteMessage(BaseModel):
     __tablename__ = "autodelete_message"
 
-    user_id: Mapped[int] = mapped_column(Integer)
+    user_id: Mapped[int] = mapped_column(BigInteger)
     message_id: Mapped[int] = mapped_column(Integer)
     current_time: Mapped[int] = mapped_column(Integer)
