+++ infrastructure/database/models/banned_users.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Text, ForeignKey
+from sqlalchemy import Integer, Text, ForeignKey, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import BaseModel
@@ -7,5 +7,5 @@ from .base import BaseModel
 class BannedUsers(BaseModel):
     __tablename__ = "banned_users"
 
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"))
+    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id"))
     ban_reason: Mapped[str] = mapped_column(Text)
