+++ infrastructure/database/models/bid_and_sale_user_symbols.py
@@ -1,5 +1,5 @@
 from .base import BaseModel
-from sqlalchemy import Integer, ForeignKey
+from sqlalchemy import Integer, ForeignKey, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 
@@ -7,7 +7,7 @@ class BidAndSaleUserSymbols(BaseModel):
     __tablename__ = "bid_and_sale_user_symbols"
 
     user_id: Mapped[int] = mapped_column(
-        Integer, ForeignKey("bid_and_sale_users.user_id")
+        BigInteger, ForeignKey("bid_and_sale_users.user_id")
     )
     symbol_id: Mapped[int] = mapped_column(
         Integer, ForeignKey("bid_and_sale_symbols.id")
