+++ infrastructure/database/models/closed_leaderboard_orders.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Numeric, String, Boolean, Integer
+from sqlalchemy import Numeric, String, Boolean, Integer, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import BaseModel
@@ -8,7 +8,7 @@ class ClosedLeaderboardOrders(BaseModel):
     __tablename__ = "closed_leaderboard_orders"
 
     position_id: Mapped[int] = mapped_column(Integer)
-    user_id: Mapped[int] = mapped_column(Integer)
+    user_id: Mapped[int] = mapped_column(BigInteger)
     amount: Mapped[float] = mapped_column(Numeric)
     symbol: Mapped[str] = mapped_column(String)
     entry_price: Mapped[float] = mapped_column(Numeric)
