+++ infrastructure/database/models/forex_users.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Boolean
+from sqlalchemy import Integer, Boolean, BigInteger
 from sqlalchemy.orm import mapped_column, Mapped
 
 from .base import BaseModel
@@ -7,5 +7,5 @@ from .base import BaseModel
 class ForexUsers(BaseModel):
     __tablename__ = "forex_users"
 
-    user_id: Mapped[int] = mapped_column(Integer)
+    user_id: Mapped[int] = mapped_column(BigInteger)
     is_active: Mapped[bool] = mapped_column(Boolean, default=False)
