+++ infrastructure/database/models/klines_pairs.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, Numeric
+from sqlalchemy import Integer, String, Numeric, BigInteger
 from sqlalchemy.orm import mapped_column, Mapped
 
 from .base import BaseModel
@@ -7,7 +7,7 @@ from .base import BaseModel
 class KlinesPairs(BaseModel):
     __tablename__ = "klines_pairs"
 
-    user_id: Mapped[int] = mapped_column(Integer)
+    user_id: Mapped[int] = mapped_column(BigInteger)
     pair: Mapped[str] = mapped_column(String)
     bid_value: Mapped[float] = mapped_column(Numeric, default=500000)
     sale_value: Mapped[float] = mapped_column(Numeric, default=500000)
