+++ infrastructure/database/models/klines_users.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Boolean, Float
+from sqlalchemy import Integer, Boolean, Float, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import BaseModel
@@ -7,7 +7,7 @@ from .base import BaseModel
 class KlinesUsers(BaseModel):
     __tablename__ = "klines_users"
 
-    user_id: Mapped[int] = mapped_column(Integer)
+    user_id: Mapped[int] = mapped_column(BigInteger)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     short_value: Mapped[float] = mapped_column(Float, default=0.0)
     long_value: Mapped[float] = mapped_column(Float, default=0.0)
