+++ infrastructure/database/models/leaderboard_orders.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Boolean
+from sqlalchemy import Integer, Boolean, BigInteger
 from sqlalchemy.orm import mapped_column, Mapped
 
 from .base import BaseModel
@@ -8,5 +8,5 @@ class LeaderboardOrders(BaseModel):
     __tablename__ = "leaderboard_orders"
 
     position_id: Mapped[int] = mapped_column(Integer)
-    user_id: Mapped[int] = mapped_column(Integer)
+    user_id: Mapped[int] = mapped_column(BigInteger)
     is_posted: Mapped[bool] = mapped_column(Boolean, default=False)
