+++ infrastructure/database/models/referral_activities.py
@@ -1,5 +1,5 @@
 from .base import BaseModel
-from sqlalchemy import Integer, Text, ForeignKey
+from sqlalchemy import Integer, Text, ForeignKey, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 
@@ -7,7 +7,7 @@ class ReferralActivities(BaseModel):
     __tablename__ = "referral_activities"
 
     user_id: Mapped[int] = mapped_column(
-        Integer, ForeignKey("user_referrals.referral_user_id")
+        BigInteger, ForeignKey("user_referrals.referral_user_id")
     )
     url: Mapped[str] = mapped_column(Text)
     remote_addr: Mapped[str] = mapped_column(Text)
